--- src/test/java/nextstep/ReservationE2ETest.java
@@ -1,134 +1,0 @@
-package nextstep;
-
-import io.restassured.RestAssured;
-import nextstep.reservation.Reservation;
-import nextstep.reservation.ReservationRequest;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.test.annotation.DirtiesContext;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-class ReservationE2ETest {
-    public static final String DATE = "2022-08-11";
-    public static final String TIME = "13:00";
-    public static final String NAME = "name";
-
-    private ReservationRequest request;
-
-    @BeforeEach
-    void setUp() {
-        request = new ReservationRequest(
-                DATE,
-                TIME,
-                NAME
-        );
-    }
-
-    @DisplayName("예약을 생성한다")
-    @Test
-    void create() {
-        var response = RestAssured
-                .given().log().all()
-                .body(request)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/reservations")
-                .then().log().all()
-                .extract();
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-    }
-
-    @DisplayName("예약을 조회한다")
-    @Test
-    void show() {
-        createReservation();
-
-        var response = RestAssured
-                .given().log().all()
-                .param("date", DATE)
-                .when().get("/reservations")
-                .then().log().all()
-                .extract();
-
-        List<Reservation> reservations = response.jsonPath().getList(".", Reservation.class);
-        assertThat(reservations.size()).isEqualTo(1);
-    }
-
-    @DisplayName("예약을 삭제한다")
-    @Test
-    void delete() {
-        createReservation();
-
-        var response = RestAssured
-                .given().log().all()
-                .param("date", DATE)
-                .param("time", TIME)
-                .when().delete("/reservations")
-                .then().log().all()
-                .extract();
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
-
-    @DisplayName("중복 예약을 생성한다")
-    @Test
-    void createDuplicateReservation() {
-        createReservation();
-
-        var response = RestAssured
-                .given().log().all()
-                .body(request)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/reservations")
-                .then().log().all()
-                .extract();
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    @DisplayName("예약이 없을 때 예약 목록을 조회한다")
-    @Test
-    void showEmptyReservations() {
-        var response = RestAssured
-                .given().log().all()
-                .param("date", DATE)
-                .when().get("/reservations")
-                .then().log().all()
-                .extract();
-
-        List<Reservation> reservations = response.jsonPath().getList(".", Reservation.class);
-        assertThat(reservations.size()).isEqualTo(0);
-    }
-
-    @DisplayName("없는 예약을 삭제한다")
-    @Test
-    void createNotExistReservation() {
-        var response = RestAssured
-                .given().log().all()
-                .param("date", DATE)
-                .param("time", TIME)
-                .when().delete("/reservations")
-                .then().log().all()
-                .extract();
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    private void createReservation() {
-        RestAssured
-                .given().log().all()
-                .body(request)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/reservations")
-                .then().log().all();
-    }
-}
